Restructuring serial code
Restructuring the serial code to behave a bit better.  The old code
was dysfunctional because it would wait forever for reads that
would never come.  This version uses the fact that the frames
being sent to us always have a length.  We use that to be able
to read the proper number of bytes for the frame.  It's debatable
whether we need some of this complexity because unlike a hardware
serial port, a pty should be fairly reliable.

Also took out code for setting serial parameters on a pty.  They
are all useless noops.  Since we *know* we'll be using a pty,
don't even try to set them.  Also move to using posix_openpt()
instead of keeping what the pty master is going to be.  Let the
OS do its job.

Made parsing input better by doing so with packed structs.  This
should eliminate a lot of opaque byte counting and size
calculations.